+++ spec/silas/cancellation_spec.rb
@@ -20,7 +20,7 @@ RSpec.describe "turn cancellation" do
 
   def configure!(engine, tool)
     Silas.configure do |c|
-      c.engine = engine
+      c.adapter = engine
       c.isolate_steps = false
       c.tool_resolver = ->(_name) { tool }
     end